Get a grid of cells rendering
This is a good step towards cellular automata, but it may be a dead end.
What I want to do next is define a block cellular automaton, but that
requires operations on the grid that are only well-defined if it's of
even size, and the type system of Elm isn't strong enough to make or
understand those guarantees.

I'm going to have a look at other languages and see if there's something
else that can express this better.